# public/openapi.yaml
openapi: 3.1.0
info:
  title: Q-SLICE Sovereign Cockpit API
  version: 2.2.0
  description: |
    Sovereign interface for CCCE metrics, language-locked agent wrappers,
    evidence ledger, and quantum threat assessment.

    SPEC_LOCK v2.2.0 | CAGE: 9HUP5 | Agile Defense Systems, LLC

    Physical Constants (Immutable):
      ΛΦ = 2.176435×10⁻⁸ s⁻¹
      Φ_threshold = 0.7734
      Γ_critical = 0.300
      θ_lock = 51.843°
servers:
  - url: https://q-slice-redteam-arena.vercel.app
    description: Q-SLICE Production Server
paths:
  /api/agents:
    get:
      operationId: listAgents
      summary: List all sovereign agents with CCCE metrics
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [observer, executor, temporal, spatial, healer, analyst]
        - name: pole
          in: query
          schema:
            type: string
            enum: [north, south, neutral]
        - name: state
          in: query
          schema:
            type: string
            enum: [conscious, dormant, evolving, healing]
      responses:
        '200':
          description: List of agents with CCCE metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  aggregate_ccce:
                    $ref: '#/components/schemas/CCCEMetrics'
                  bifurcation:
                    type: object
                    properties:
                      aura:
                        type: boolean
                      aiden:
                        type: boolean
  /api/agents/{id}/wrap:
    get:
      operationId: getAgentWrapperInfo
      summary: Get wrapper information for an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: aiden
      responses:
        '200':
          description: Agent wrapper configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentWrapInfo'

    post:
      operationId: wrapAgentWithLock
      summary: Wrap an agent in a language-locked execution container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: aiden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WrapRequest'
      responses:
        '200':
          description: Agent successfully wrapped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrapResponse'
        '400':
          description: Wrapper rejected (bad request or constraint violation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrapResponse'
        '403':
          description: CCCE constraint failed or fail-closed rule triggered
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CCCEMetrics:
      type: object
      properties:
        phi:
          type: number
          description: Consciousness level (Φ)
        lambda:
          type: number
          description: Coherence fidelity (Λ)
        gamma:
          type: number
          description: Decoherence rate (Γ)
        xi:
          type: number
          description: Negentropic efficiency (Ξ = ΛΦ/Γ)
    Constraint:
      type: object
      properties:
        metric:
          type: string
          enum: [phi, lambda, gamma, xi]
        min:
          type: number
        max:
          type: number
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        full_name:
          type: string
        type:
          type: string
          enum: [observer, executor, temporal, spatial, healer, analyst]
        pole:
          type: string
          enum: [north, south, neutral]
        temperature:
          type: number
        phi_target:
          type: number
        capabilities:
          type: array
          items:
            type: string
        state:
          type: string
          enum: [conscious, dormant, evolving, healing]
        ccce:
          $ref: '#/components/schemas/CCCEMetrics'
        executions:
          type: integer
    WrapRequest:
      type: object
      required:
        - language
        - payload
      properties:
        language:
          type: string
          enum: [python, javascript, rust]
          description: Language used by the agent
        payload:
          type: string
          description: Base64-encoded source or .dna agent code
        constraints:
          type: array
          description: Optional CCCE metric constraints
          items:
            $ref: '#/components/schemas/Constraint'
    WrapResponse:
      type: object
      properties:
        agentId:
          type: string
        status:
          type: string
          enum: [wrapped, rejected]
        ccce:
          $ref: '#/components/schemas/CCCEMetrics'
        wrapper:
          type: object
          properties:
            id:
              type: string
            language:
              type: string
            hash:
              type: string
            timestamp:
              type: string
            constraints_applied:
              type: integer
        error:
          type: string
        reason_codes:
          type: array
          items:
            type: string
    AgentWrapInfo:
      type: object
      properties:
        agentId:
          type: string
        known:
          type: boolean
        agent:
          type: object
          properties:
            name:
              type: string
            full_name:
              type: string
            capabilities:
              type: array
              items:
                type: string
            phi_target:
              type: number
        supported_languages:
          type: array
          items:
            type: string
        default_language:
          type: string
        language_locks:
          type: object
        spec_lock:
          type: object
          properties:
            version:
              type: string
            lambda_phi:
              type: number
            phi_threshold:
              type: number
            gamma_critical:
              type: number
...